Trocando tipo

+++ include/digiapi.h
@@ -3,9 +3,9 @@
 
 #include "digimon.h"
 
-int DIGI_init(const char* szSaveFile);
+uint8_t DIGI_init(const char* szSaveFile);
 
-int DIGI_initDigitama(const char* szSaveFile, uint8_t uiDigitamaIndex);
+uint8_t DIGI_initDigitama(const char* szSaveFile, uint8_t uiDigitamaIndex);
 
 uint8_t DIGI_updateEventsDeltaTime(uint16_t uiDeltaTime, uint8_t* puiEvents);
